create layout.tsx for balance of payment dashboard and rename files
